+++ sql/types/public_get_currencies.sql
@@ -1,7 +1,7 @@
 drop type if exists deribit.public_get_currencies_response_withdrawal_priority cascade;
 create type deribit.public_get_currencies_response_withdrawal_priority as (
 	name text,
-	value float
+	value double precision
 );
 
 
@@ -12,8 +12,8 @@ create type deribit.public_get_currencies_response_result as (
 	currency_long text,
 	fee_precision bigint,
 	min_confirmations bigint,
-	min_withdrawal_fee float,
-	withdrawal_fee float,
+	min_withdrawal_fee double precision,
+	withdrawal_fee double precision,
 	withdrawal_priorities deribit.public_get_currencies_response_withdrawal_priority[]
 );
 comment on column deribit.public_get_currencies_response_result.coin_type is 'The type of the currency.';
